Added 'shadow', 'emit_rad' and 'recv_rad' parameters to Wings material editor, to be used on YafRay objects from the default material.
This should bring the YafRay exporter to the feature level of
YafRay version 0.0.3, which is the latest "stable" release,
with the exception of Arealights and the experimental 'indirect_samples'
rendering parameters.

However, the exporter requires YafRay CVS snapshot 20030309 to work
due to necessary bug fixes and features in the renderer, such as
the negative UV coordinates bug and .tga texture files.

But, the exporter does not support many new features in the CVS
snapshot. These new features and Arealights are on the todo list.

Thereby I dare declare the YafRay exporter to now be a beta release.

/ Raimo Niskanen (ratmapper)