wings_ask.erl: Made it possible for a button with action 'keep' to return a new variable store.
wpc_yafray.erl:
Rewrote the YafRay options in the material editor to give
better control of YafRay specific colors. The mapping of OpenGL colors
to YafRay colors has changed slightly. Fresnel parameters are now collected
in a subframe and grazing angle colors has been implemented.

The default values for an untouched material has not changed. When editing
a material, the old value for "Transmitted", i.e Diffuse*(1-Opacity),
can now be set with a button. The old value for "Reflected", i.e Ambient,
can be set by drag-and-drop from Ambient. To map Ambient to "Reflected"
was not really a good idea anyway, and the old mapping of "Transmitted"
was only designed to make semi-transparent objects in OpenGL look the
same in YafRay (if Index Of Refraction was 1.0).