e3d__tri_quad.erl: Fixed to propagate vertex colors during triangulation.
e3d_mesh.erl: Implemented slit_hard_edges/1,2 to operate on #e3d_mesh{}.

wpc_yafray.erl: Implemented simulated hard edges by slitting the mesh.
Introduced a checkbox "Use Edge Hardness" to activate this feature from
the material editor. Created a new subframe in the material editor for
"Object Parameters". These are the options that affect the whole object
that has the material on the majority of its faces. Note that there are
some hidden object parameters: if you activate the "Caustic" object
parameter the fresnel parameters "IOR", "Reflected" and "Transmitted" are
transformed to their caustic correnspondents and set on the object.