Bug fix: The value of a menu field was not updated if you exited with the space key unless the field had got mouse focus, so if you tabbed to the field it did not work.
Introduced checkboxed minimized h/v-frames. Rather quick and dirty. It is
just h/v-frames with a different look that suggests to the user that the
minimized state has got significance, so therefore the minimized state
return value can be used (as the negative checkbox value, and it is also
to get the frame to be minimized when the checkbox is checked). Use
the flags 'checkbox' same as {checkbox,true}, or {checkbox,false} for the
latter reversed variant.