wings_help.erl: Exported cmd/1.
wings_ask.erl: Implemented the notion of horizontal stretch. Now a panel
field (default stretch 1) can be used to get fields to the right of the
panel field in a hframe at the right border.

wpc_yafray.erl: Used the new horizontal stretch property on panels to move
the help butons to the right border instead of the left.